Add gst-plugins-bad
GStreamer Bad Plug-ins is a set of plug-ins that aren't up to par compared to
the rest. They might be close to being good quality, but they're missing
something - be it a good code review, some documentation, a set of tests, a
real live maintainer, or some actual wide use.

http://gstreamer.freedesktop.org/modules/gst-plugins-bad.html