(FL-111) Inlcude jokosher in Foresight
Add Gnonlin which is a requirement of Jokosher.

Gnonlin is a set of GStreamer elements to ease the creation of non-linear
multimedia editors. It works together with the GStreamer multimedia framework
to give developers a powerfull and flexible set of tools for quickly assembling
applications which needs to handle non-linear multimedia editing.

WWW: http://gnonlin.sourceforge.net