Re-enable the native hint interpreter.
Enabling this by itself doesn't have the visual impact of the cleartype patches
we had originally testing with this. In fact for most fonts it's hard to tell
the difference. Certain glyphs do benefit from it being enabled though - my
testcase being the dotted zero in Vera Mono (yuck).

I also removed the 'fluff' from the recipe. Since its either on or off, I don't
see the point of setting meaningless flags in the recipe.