Add emerald.
The Emerald Window Decorator is a custom window decorator shipped with Compiz
Fusion that allows for theming and full composite window decorations with the
use of engines. Emerald allows for all sorts of different configurations and
layouts of buttons, look, title bars and frames. Emerald Window Decorator is
completely independent of any desktop, and you specify your own themes to use
for it.

WWW: http://wiki.compiz-fusion.org/Decorators/Emerald