http://www.opensc-project.org/openct/
OpenCT implements drivers for several smart card readers.
It comes as driver in ifdhandler format for PC/SC-Lite, as CT-API driver, or as a small and lean middleware,
so applications can use it with minimal overhead. OpenCT also has a primitive mechanism to export smart card
readers to remote machines via tcp/ip.